requirements.txt
langchain==0.2.17
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-community==0.2.17
langchain-text-splitters==0.2.4
faiss-cpu==1.7.4
openai>=1.56.1
httpx==0.27.2
tiktoken==0.8.0
gradio>=5.0.0          
jinja2>=3.1.4          
python-dotenv==1.0.1
huggingface-hub==0.25.2
